requirements.txt
# Virtual Saree Try-On Backend Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# AI/ML Libraries (for production deployment)
# torch==2.1.0
# torchvision==0.16.0
# onnxruntime==1.16.1
# mediapipe==0.10.7

# Background Removal
# rembg==2.0.50

# Super Resolution
# realesrgan==0.3.0

# Pose Detection
# ultralytics==8.0.196

# Utilities
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0

# Optional: Database (if needed)
# sqlalchemy==2.0.23
# alembic==1.12.1

# Optional: Redis (for caching)
# redis==5.0.1

# Optional: Monitoring
# prometheus-client==0.18.0